refactor(tag): move tag-related classes into dedicated package
