+++ tracker-api/src/main/java/com/sloyardms/trackerapi/tag/TagController.java
@@ -1,10 +1,9 @@
-package com.sloyardms.trackerapi.controller;
+package com.sloyardms.trackerapi.tag;
 
-import com.sloyardms.trackerapi.dto.TagCreateDto;
-import com.sloyardms.trackerapi.dto.TagDto;
-import com.sloyardms.trackerapi.dto.TagUpdateDto;
+import com.sloyardms.trackerapi.tag.dto.TagCreateDto;
+import com.sloyardms.trackerapi.tag.dto.TagDto;
+import com.sloyardms.trackerapi.tag.dto.TagUpdateDto;
 import com.sloyardms.trackerapi.security.AuthUtils;
-import com.sloyardms.trackerapi.service.interfaces.TagService;
 import jakarta.validation.Valid;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
