+++ tracker-api/src/main/java/com/sloyardms/trackerapi/tag/dto/TagUpdateDto.java
@@ -1,18 +1,16 @@
-package com.sloyardms.trackerapi.dto;
+package com.sloyardms.trackerapi.tag.dto;
 
-import jakarta.validation.constraints.NotBlank;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-@Getter
 @Setter
+@Getter
 @NoArgsConstructor
 @AllArgsConstructor
-public class TagCreateDto {
+public class TagUpdateDto {
 
-    @NotBlank
     private String name;
 
 }
