+++ tracker-api/src/main/java/com/sloyardms/trackerapi/tag/entity/Tag.java
@@ -1,5 +1,6 @@
-package com.sloyardms.trackerapi.entity;
+package com.sloyardms.trackerapi.tag.entity;
 
+import com.sloyardms.trackerapi.entity.Auditable;
 import com.sloyardms.trackerapi.user.entity.User;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
